+++ src/main/kotlin/dev/nyon/skylper/skyblock/Mining.kt
@@ -1,9 +1,0 @@
-package dev.nyon.skylper.skyblock
-
-import dev.nyon.skylper.skyblock.mining.hollows.HollowsModule
-
-object Mining {
-    fun init() {
-        HollowsModule.init()
-    }
-}